requirements.txt
pytesseract 
pillow 
flask
PyPDF2 
pdf2image
openai 
langchain 
pandas
langchain_community
openpyxl
sentence_transformers
faiss-cpu
mysql-connector-python
sqlalchemy
pymysql 
requests 
pillow
flask-login
langchain_openai